code/requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiogoogle==5.6.0
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
aws-cdk-lib==2.92.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
bidict==0.22.1
cachetools==5.3.2
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
constructs==10.2.69
dictdiffer==0.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
google-auth==2.23.4
idna==3.4
importlib-resources==6.0.1
iso8601==0.1.16
Jinja2==3.1.2
jsii==1.87.0
mangum==0.17.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.4
openpyxl==3.1.2
pandas==2.0.3
publication==0.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic-core==2.4.0
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-engineio==4.6.1
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.34.0
tomlkit==0.12.1
tonyg-rfc3339==0.1
tortoise-orm==0.18.1
typeguard==2.13.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
XlsxWriter==3.1.6
yarl==1.9.2
zipp==3.16.2
